a bit of copy editing
A bit few typos, and some suggested text. The substantial bit is that a duck array *could* choose to implement __array__, if it wants to coerce itself and return a proper numpy array. This could helpful, because only the object itself knows how best to make itself into an numpy array.